refactor: remove unused seedStyles function and streamline database seeding process; update prompts with unique IDs and ensure defaults are maintained
