+++ src/main/db/seed.ts
@@ -1,8 +1,7 @@
 import { app } from 'electron'
 import { eq } from 'drizzle-orm'
 import { getDb } from './index'
-import { appSettings, prompts, styles } from './schema'
-import { seedPrompts as baseSeedPrompts } from '../prompts/seed'
+import { appSettings, prompts } from './schema'
 import { randomUUID } from 'crypto'
 
 // Wrap existing seeders with idempotent ensure-logic and version tracking
@@ -14,8 +13,6 @@ export function runDataMigrations(): void {
   // Read last applied app version
   const last = db.select().from(appSettings).where(eq(appSettings.key, 'app.version')).all()
 
-  // Always ensure defaults (safe, idempotent)
-  ensureDefaultStyles()
   ensureDefaultPrompts()
 
   // Future: add versioned data migrations here
@@ -36,27 +33,25 @@ export function runDataMigrations(): void {
 
 function ensureDefaultPrompts(): void {
   const db = getDb()
-  // Run base seeder (inserts if empty)
-  baseSeedPrompts()
 
   // For shipped prompt ids, upsert and refresh content on upgrades
   const DEFAULTS = [
     {
-      id: 'outpainting',
+      id: cryptoRandomUUID(),
       name: '智能扩图',
       description: '基于原图内容智能扩展画面，保持原始风格和构图',
       prompt: 'expand and extend the image naturally while maintaining the original style and composition',
       category: 'outpainting',
     },
     {
-      id: 'enhance',
+      id: cryptoRandomUUID(),
       name: '图像增强',
       description: '提升图像质量，增强细节和清晰度',
       prompt: 'enhance image quality, improve details and sharpness while preserving the original content',
       category: 'enhancement',
     },
     {
-      id: 'colorize',
+      id: cryptoRandomUUID(),
       name: '智能上色',
       description: '为黑白图像添加自然色彩',
       prompt: 'add natural colors to black and white image, maintain realistic tones',
@@ -69,7 +64,6 @@ function ensureDefaultPrompts(): void {
     if (found.length === 0) {
       db.insert(prompts).values(p).run()
     } else {
-      // Update shipped defaults to latest content (user can add their own rows separately)
       db
         .update(prompts)
         .set({ name: p.name, description: p.description, prompt: p.prompt, category: p.category })
@@ -79,4 +73,13 @@ function ensureDefaultPrompts(): void {
   }
 }
 
-// no-op helper removed; using randomUUID directly
+function cryptoRandomUUID() {
+  if (typeof globalThis.crypto?.randomUUID === 'function') return globalThis.crypto.randomUUID()
+  // Fallback simple UUIDv4 (not RFC-4122 strict, but fine for local seeds)
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, c => {
+    const r = (Math.random() * 16) | 0
+    const v = c === 'x' ? r : (r & 0x3) | 0x8
+    return v.toString(16)
+  })
+}
+
